Introduce a new VirtualPath type, "ephemeral".
An ephemeral path has a name and contents. It's the responsibility of an
argument redolver to actually write the given contents to a file with the
specified name somewhere in the file system.

For now, I've sued this to remove the FileSystem from the ExplicitModuleBuildHandler.
Going forward, it should help simplify response file and filelist support as well.